+++ packages/react-components/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `WithWalletConnector`: Removed method `connectActive` (and child prop `isConnecting`).
     Use the `connect` method directly on `WalletConnector` instead.
 -   Extract hook `useConnection` from `activeConnection` and related fields which were removed from `WithWalletConnector`.
+-   `WithWalletConnector`: The field `connectedAccounts` will map connections to the empty string if they don't have an associated account.
 
 ## [0.1.0] - 2023-01-17